refactoring + unit tests
