+++ internal/app/repository/repository.go
@@ -1,4 +1,4 @@
-package storage
+package repository
 
 var Repository = make(map[int]string)
