+++ internal/app/repository/repository_test.go
@@ -1,4 +1,4 @@
-package storage
+package repository
 
 import (
 	"github.com/stretchr/testify/assert"
